refactor: separating logic from services
