+++ src/Application/Bootstrap.cs
@@ -4,7 +4,8 @@ using Application.Behaviours;
 using Microsoft.Extensions.DependencyInjection;
 using Domain.Services;
 using Domain.Settings;
-using Application.Services;
+using Application.Services.Jwt;
+using Application.Services.Hasher;
 using Microsoft.AspNetCore.Identity;
 using Domain.Entities;
 using Microsoft.Extensions.Configuration;
